backend/requirements.txt
# Core dependencies for RAG chatbot
langchain==0.0.352
langchain-community==0.0.38
langchain-openai==0.0.8
transformers==4.35.0
torch==2.1.0
faiss-cpu==1.7.4
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
numpy==1.24.3
pandas==2.1.3
tiktoken==0.5.1
openai==1.3.5
beautifulsoup4==4.12.2
PyYAML==6.0.1